paycharm/requirements.txt
fastapi
uvicorn[standard]

SQLAlchemy
psycopg2-binary

alembic

pydantic
pydantic-settings
python-dotenv

aiogram

gspread
google-auth

openai

email-validator

python-dateutil
